Change Node class getters to return constants
Change getter functions in Node.hpp to return constant shared_ptr's of
constant type.

The values returned from the function should not be open to changing,
regardless if it is the pointer or the object the pointer points to. The
functions' only goal is to return what is held in the member variables.
There are methods designated for updating values within the node.