Change Node copy constructor to accept const
Change Node.hpp copy constructor to accept a constant shared_ptr of type
const Node<T>.

This was missed in the previous commit. Functions that accept arguments
but do not change the argument values should only accept constant values
so as not to accidentally allow the arguments to be changed. The header
file now reflects that.