Add throw condition for Node::getPrev()
Add condition for getPrev() to throw an exception when there is no
previous node.

The function throws this error because there are composite classes that
rely on Node objects to not be able to point to previous nodes (such as
Singly Linked Lists). This allows for the same Node class to be used as
long as the prev member variable points to nullptr in those classes. By
throwing an error, the class cannot attempt to access the previous node.